ConfigInputChangeEvent <-> ConfigInputChangeEventFlow.
