+++ core/cloudflow-operator-actions/src/main/scala/cloudflow/operator/event/AppEvent.scala
@@ -21,6 +21,7 @@ import scala.collection.immutable.Seq
 
 import skuber._
 import skuber.api.client._
+
 import cloudflow.operator.action._
 
 /**
