+++ core/cloudflow-operator-actions/src/main/scala/cloudflow/operator/event/ConfigInputChangeEvent.scala
@@ -22,8 +22,13 @@ import java.nio.charset.StandardCharsets
 import scala.util.Try
 import akka.actor._
 import com.typesafe.config._
+
+import org.slf4j.LoggerFactory
+
 import skuber._
 import skuber.api.client._
+import skuber.json.format._
+
 import cloudflow.operator.action._
 import cloudflow.blueprint.deployment.StreamletDeployment
 
@@ -33,6 +38,80 @@ import cloudflow.blueprint.deployment.StreamletDeployment
 case class ConfigInputChangeEvent(appId: String, namespace: String, watchEvent: WatchEvent[Secret]) extends AppChangeEvent[Secret]
 
 object ConfigInputChangeEvent extends Event {
+  val log = LoggerFactory.getLogger(this.getClass)
+
+  def toConfigInputChangeEvent(
+      currentSecrets: Map[String, WatchEvent[Secret]],
+      watchEvent: WatchEvent[Secret]
+  ): (Map[String, WatchEvent[Secret]], List[ConfigInputChangeEvent]) = {
+    val secret       = watchEvent._object
+    val metadata     = secret.metadata
+    val secretName   = secret.metadata.name
+    val namespace    = secret.metadata.namespace
+    val absoluteName = s"$namespace.$secretName"
+
+    def hasChanged(existingEvent: WatchEvent[Secret]) =
+      watchEvent._object.resourceVersion != existingEvent._object.resourceVersion && getData(existingEvent._object) != getData(secret)
+
+    watchEvent._type match {
+      case EventType.DELETED ⇒
+        (currentSecrets - absoluteName, List())
+      case EventType.ADDED | EventType.MODIFIED ⇒
+        if (currentSecrets.get(absoluteName).forall(hasChanged)) {
+          (for {
+            appId        ← metadata.labels.get(CloudflowLabels.AppIdLabel)
+            configFormat <- metadata.labels.get(CloudflowLabels.ConfigFormat) if configFormat == CloudflowLabels.InputConfig
+            _ = log.info(s"[app: $appId application configuration changed ${changeInfo(watchEvent)}]")
+          } yield {
+            (currentSecrets + (absoluteName -> watchEvent), List(ConfigInputChangeEvent(appId, namespace, watchEvent)))
+          }).getOrElse((currentSecrets, List()))
+
+        } else (currentSecrets, List())
+    }
+  }
+
+  def toActionList(mappedApp: Option[CloudflowApplication.CR],
+                   event: ConfigInputChangeEvent,
+                   podNamespace: String): Seq[Action[ObjectResource]] =
+    (mappedApp, event) match {
+      case (Some(app), configInputChangeEvent) ⇒
+        val appConfig = getConfigFromSecret(configInputChangeEvent)
+
+        val clusterNames = (for {
+            d          <- app.spec.deployments
+            (_, topic) <- d.portMappings
+            cluster    <- topic.cluster.toVector
+          } yield cluster) :+ TopicActions.DefaultConfigurationName
+
+        val providedAction = Action.providedByLabel[Secret, Secret](
+          TopicActions.KafkaClusterNameLabel,
+          clusterNames,
+          podNamespace, { clusterSecrets =>
+            val allNamedClusters = namedClusters(app.name, clusterNames, clusterSecrets)
+            val actions = app.spec.deployments.map {
+              streamletDeployment ⇒
+                val configs = ConfigurationScopeLayering.configs(streamletDeployment, appConfig, allNamedClusters)
+
+                // create update action for output secret action which is mounted as config by runtime specific deployments
+                val configSecret =
+                  createSecret(
+                    streamletDeployment.secretName,
+                    app,
+                    streamletDeployment,
+                    configs.streamlet,
+                    configs.runtime,
+                    configs.pods,
+                    CloudflowLabels.StreamletDeploymentConfigFormat
+                  )
+                Action.createOrUpdate(configSecret, secretEditor)
+            }
+
+            Action.composite(actions)
+          }
+        )
+        List(providedAction)
+      case _ ⇒ Nil // app could not be found, do nothing.
+    }
 
   val SecretDataKey        = "secret.conf"
   val RuntimeConfigDataKey = "runtime-config.conf"
@@ -55,7 +134,7 @@ object ConfigInputChangeEvent extends Event {
    * If a referenced cluster name does not have a corresponding secret K8s then log an error, but continue.
    * Return all named cluster configs and the 'default' cluster config, if one is defined.
    */
-  def namedClusters(appName: String, clusterNames: Vector[String], clusterSecrets: ListResource[Secret], system: ActorSystem) = {
+  def namedClusters(appName: String, clusterNames: Vector[String], clusterSecrets: ListResource[Secret]) = {
     val namedClusters: Map[String, Config] = clusterNames.flatMap { name =>
       val secret = clusterSecrets.items.find(_.metadata.labels.get(TopicActions.KafkaClusterNameLabel).contains(name))
       secret match {
@@ -63,7 +142,7 @@ object ConfigInputChangeEvent extends Event {
           val clusterConfigStr = getConfigFromSecret(secret)
           Vector(name -> clusterConfigStr)
         case None =>
-          system.log.error(
+          log.error(
             s"""
               |The referenced cluster configuration secret '$name' for app '$appName' does not exist.
               |This should have been detected at deploy time.
@@ -82,7 +161,7 @@ object ConfigInputChangeEvent extends Event {
     namedClusters ++ maybeDefaultCluster
   }
 
-  def getConfigFromSecret(configInputChangeEvent: ConfigInputChangeEvent, system: ActorSystem) = {
+  def getConfigFromSecret(configInputChangeEvent: ConfigInputChangeEvent) = {
     val secret: Secret = configInputChangeEvent.watchEvent._object
     secret.data
       .get(ConfigInputChangeEvent.SecretDataKey)
@@ -90,15 +169,15 @@ object ConfigInputChangeEvent extends Event {
         val str = new String(bytes, StandardCharsets.UTF_8)
         Try(ConfigFactory.parseString(str).resolve()).recover {
           case cause =>
-            system.log.error(
-              cause,
-              s"Detected input secret '${secret.metadata.name}' contains invalid configuration data, IGNORING configuration."
+            log.error(
+              s"Detected input secret '${secret.metadata.name}' contains invalid configuration data, IGNORING configuration.",
+              cause
             )
             ConfigFactory.empty()
         }.toOption
       }
       .getOrElse {
-        system.log.error(
+        log.error(
           s"Detected input secret '${secret.metadata.name}' does not have data key '${ConfigInputChangeEvent.SecretDataKey}', IGNORING configuration."
         )
         ConfigFactory.empty()
@@ -143,4 +222,5 @@ object ConfigInputChangeEvent extends Event {
   def secretEditor = new ObjectEditor[Secret] {
     def updateMetadata(obj: Secret, newMetadata: ObjectMeta): Secret = obj.copy(metadata = newMetadata)
   }
+
 }
